apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
supabase==2.3.4
python-multipart==0.0.9
alembic==1.13.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0